+++ CHANGES.rst
@@ -83,6 +83,8 @@ ibe
 imcce
 ^^^^^
 
+- Fix Skybot return for unumbered asteroids. [#1598]
+
 irsa
 ^^^^